feat: 새로운 앱 (project_detail , search 추가)
