+++ accounts/models.py
@@ -1,18 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
-
-class UniversityTag(models.Model):
-    # 대학교 이름
-    university_name = models.CharField(null=False, max_length=255)
-
-class GenreTag(models.Model):
-    # 장르 이름
-    genre_name = models.CharField(null=False, max_length=255)
-
-class StackTag(models.Model):
-    # 스택 이름
-    stack_name = models.CharField(null=False, max_length=255)
-
+from search.models import UniversityTag, GenreTag
 
 class Account(models.Model):
     # id, username, password, is_staff, is_superuser, last_login, date_joined 필드가 존재
