+++ project/urls.py
@@ -22,5 +22,7 @@ from django.conf.urls.static import static
 urlpatterns = [
     path('api/admin/', admin.site.urls),
     path('api/accounts/', include('accounts.urls')),
+    path('api/search/', include('search.urls')),
     path('api/project/', include('ptn_project.urls')),
+    path('api/project_detail/', include('project_detail.urls')),
 ]+ static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
