+++ project_detail/admin.py
@@ -1,5 +1,3 @@
 from django.contrib import admin
-from .models import Account
 
 # Register your models here.
-admin.site.register(Account)
