+++ project_detail/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+
+app_name = 'project_detail'
+urlpatterns = [
+  
+]
